CLEANUP: Move $multiplier and $sizes to ScalungConfiguration Object
Relates: #22, relates: #24
