+++ Tests/Fixtures/container_example/Configuration/TypoScript/Container/2col-33-66.typoscript
@@ -7,9 +7,9 @@ plugin.tx_responsiveimages {
                         multiplier {
                             xs = 1
                             sm = 1
-                            md = 0,333
-                            lg = 0,333
-                            xl = 0,333
+                            md = 0.333
+                            lg = 0.333
+                            xl = 0.333
                         }
                     }
 
@@ -17,9 +17,9 @@ plugin.tx_responsiveimages {
                         multiplier {
                             xs = 1
                             sm = 1
-                            md = 0,666
-                            lg = 0,666
-                            xl = 0,666
+                            md = 0.666
+                            lg = 0.666
+                            xl = 0.666
                         }
                     }
                 }
