+++ Tests/Fixtures/container_example/Configuration/TypoScript/Container/3col.typoscript
@@ -7,9 +7,9 @@ plugin.tx_responsiveimages {
                         multiplier {
                             xs = 1
                             sm = 1
-                            md = 0,333
-                            lg = 0,333
-                            xl = 0,333
+                            md = 0.333
+                            lg = 0.333
+                            xl = 0.333
                         }
                     }
 
@@ -17,9 +17,9 @@ plugin.tx_responsiveimages {
                         multiplier {
                             xs = 1
                             sm = 1
-                            md = 0,333
-                            lg = 0,333
-                            xl = 0,333
+                            md = 0.333
+                            lg = 0.333
+                            xl = 0.333
                         }
                     }
 
@@ -27,9 +27,9 @@ plugin.tx_responsiveimages {
                         multiplier {
                             xs = 1
                             sm = 1
-                            md = 0,333
-                            lg = 0,333
-                            xl = 0,333
+                            md = 0.333
+                            lg = 0.333
+                            xl = 0.333
                         }
                     }
                 }
